Fix endless redirection loop for jellyfin in subpath nginx config.
The trailing /jellyfin/ in the proxy_pass directive causes an
endless redirection loop in a web browser.

On chromium-based browsers, the error is: "This page isn't working.
<domain_name> redirected you too many times. ERR_TOO_MANY_REDIRECTS".
On firefox-based browsers, the error is: "The page isn’t redirecting
properly. The browser has stopped trying to retrieve the requested
item. The site is redirecting the request in a way that will never
complete."

I also relocated the position of some comments to avoid being misleading
as to what they reference.