Update troubleshooting.md
Add section for troubleshooting media identification. One of the most common culprits is misnamed content. Moving the content out of the library before attempting to fix the naming gives JF a bit of helping hand by allowing it to remove the troubled item(s) and identify it fresh like it was just added.